Refactor of files structure
