+++ app/src/main/AndroidManifest.xml
@@ -36,7 +36,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".GoogleMaps.MapsActivity"
+            android:name=".Search.GoogleMaps.MapsActivity"
             android:screenOrientation="portrait"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar" />
         <activity android:name=".SignUpPage"></activity>
