+++ app/src/main/java/politechnika/meetyourtrainer/MainActivity.java
@@ -8,6 +8,8 @@ import androidx.fragment.app.FragmentManager;
 
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 
+import politechnika.meetyourtrainer.Search.FragmentSearch;
+
 public class MainActivity extends AppCompatActivity {
 
     private BottomNavigationView bottomNavigationView;
