+++ app/src/main/java/politechnika/meetyourtrainer/Search/SearchListView.java
@@ -1,4 +1,4 @@
-package politechnika.meetyourtrainer;
+package politechnika.meetyourtrainer.Search;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -15,6 +15,7 @@ import org.json.JSONObject;
 import politechnika.meetyourtrainer.Profile.ProfileActivity;
 import politechnika.meetyourtrainer.Profile.ProfileProvider;
 import politechnika.meetyourtrainer.Profile.ServerCallback;
+import politechnika.meetyourtrainer.R;
 
 public class SearchListView extends Fragment {
     //SearchListViewModel slvm;
