+++ app/src/main/res/layout/fragment_map_view.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".GoogleMaps.SearchMapView">
+    tools:context=".Search.GoogleMaps.SearchMapView">
 
     <com.google.android.gms.maps.MapView
         android:layout_height="match_parent"
